Rename dashboard folder to match component used in router
