+++ src/containers/dashboard/MapInfoDistrict.js
