+++ src/containers/dashboard/parsingmethods/getMapData.js
